sdk controller for single axios.create
this way we're not creating axios instances withe very request we make from every single class instance we create. This can really blow up

I'm trying to reduce it further and see if we can have a shared axios instance between all classes but i'm not sure this is possible.

It might be even more efficient not to create an axios instance at all and just use the default instance and create methods in SDK Controller for get, put, delete, and whatever other axios methods we use.

if we pull a world with 1,000 assets this will also create 1,000 axios instances. All of which are identical. pretty wasteful.